chore: optimize

+++ android/src/main/kotlin/io/agora/agora_rtc_engine/AgoraSurfaceViewFactory.kt
@@ -83,7 +83,7 @@ class AgoraSurfaceView(
 
   private fun setData(data: Map<*, *>) {
     val channel = (data["channelId"] as? String)?.let { getChannel(it) }
-    getEngine()?.let { view.setData(it, channel, (data["uid"] as Number).toInt()) }
+    getEngine()?.let { view.setData(it, channel, data["uid"] as Number) }
   }
 
   private fun setRenderMode(renderMode: Int) {
